requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
certifi==2023.11.17
decorator==5.1.1
distro==1.9.0
Django==5.0.1
django-filter==23.5
djangorestframework==3.14.0
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.6.1
executing==2.0.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
inflection==0.5.1
ipython==8.21.0
jedi==0.19.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
Markdown==3.5.2
matplotlib-inline==0.1.6
nanoid==2.0.0
openai==1.10.0
packaging==24.1
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stack-data==0.6.3
tqdm==4.66.1
traitlets==5.14.1
typing_extensions==4.9.0
uritemplate==4.1.1
wcwidth==0.2.13
